+++ crates/zng-task/src/http.rs
@@ -925,7 +925,7 @@ where
 {
     let mut ci = CLIENT_INIT.lock();
     if let ClientInit::Inited = &*ci {
-        Err(DefaultAlreadyInitedError)
+        Err(DefaultAlreadyInitedError {})
     } else {
         *ci = ClientInit::Set(Box::new(init));
         Ok(())
@@ -935,7 +935,7 @@ where
 /// Error returned by [`set_default_client_init`] if the default was already initialized.
 #[derive(Debug, Clone, Copy)]
 #[non_exhaustive]
-pub struct DefaultAlreadyInitedError;
+pub struct DefaultAlreadyInitedError {}
 impl fmt::Display for DefaultAlreadyInitedError {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         write!(f, "default client already initialized, can only set before first use")
@@ -1779,7 +1779,7 @@ impl fmt::Display for Error {
                 content_length,
                 max_length,
             } => write!(f, "{}", MaxLengthError(*content_length, *max_length)),
-            Error::RequireLength => write!(f, "{RequireLengthError}"),
+            Error::RequireLength => write!(f, "{}", RequireLengthError {}),
         }
     }
 }
@@ -1800,7 +1800,8 @@ impl fmt::Display for MaxLengthError {
 impl StdError for MaxLengthError {}
 
 #[derive(Debug)]
-struct RequireLengthError;
+#[non_exhaustive]
+struct RequireLengthError {}
 impl fmt::Display for RequireLengthError {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         write!(f, "content-length is required")