+++ tests/render-tests/src/tests.rs
@@ -34,10 +34,10 @@ pub async fn bw_rgb(render_mode: RenderMode, scale_factor: Factor) {
     while img.with(Img::is_loading) {
         if task::with_deadline(img.wait_update(), 20.secs()).await.is_err() {
             panic!(
-                "img wait_update timeout after 20s, img.is_loading: {}, APP.is_running: {}, VIEW_PROCESS.is_online: {}, VIEW_PROCESS ping: {}",
+                "img wait_update timeout after 20s, img.is_loading: {}, APP.is_running: {}, VIEW_PROCESS.is_connected: {}, VIEW_PROCESS ping: {}",
                 img.with(Img::is_loading),
                 APP.is_running(),
-                zng_app::view_process::VIEW_PROCESS.is_online(),
+                zng_app::view_process::VIEW_PROCESS.is_connected(),
                 zng_app::view_process::VIEW_PROCESS.image_decoders().is_ok(),
             );
         }